genai-assignment-4/requirements.txt
google-adk
chromadb
dotenv
PyMuPDF # fitz is the import name for PyMuPDF
openai # For embeddings, if OpenAIEmbeddingFunction is used 